add state so graphs re-render
