+++ src/dataset-builder/CohortEditor.ts
@@ -19,6 +19,13 @@ import {
   ProgramDataRangeCriteria,
 } from 'src/dataset-builder/DatasetBuilderUtils';
 import {
+  cohortAgeData,
+  cohortDemographicData,
+  CohortDemographics,
+  generateAgeSeries,
+  generateDemographicSeries,
+} from 'src/dataset-builder/TestConstants';
+import {
   DataRepo,
   SnapshotBuilderCountResponse,
   SnapshotBuilderDomainOption,
@@ -581,78 +588,7 @@ interface CohortEditorProps {
   readonly addSelectedCohort: (cohort: Cohort) => void;
   readonly getNextCriteriaIndex: () => number;
 }
-interface ChartLabel {
-  short: string;
-  long?: string;
-}
-interface ChartSeries {
-  name?: string;
-  data: number[];
-}
 
-interface CohortDemographics {
-  categories: ChartLabel[];
-  series: ChartSeries[];
-  title: string;
-  yTitle: string;
-  height: string;
-  legendEnabled: boolean;
-  showSeriesName: boolean;
-}
-
-function generateRandomNumbers(numNumbers: number, max: number) {
-  const randomNumbers: number[] = [];
-  for (let i = 0; i < numNumbers; i++) {
-    const randomNumber = Math.floor(Math.random() * max) + 1;
-    randomNumbers.push(randomNumber);
-  }
-  return randomNumbers;
-}
-
-function generateRandomNumbersThatAddUpTo(total: number, numNumbers: number): number[] {
-  const randomNumbers: number[] = [];
-  let remaining = total;
-  for (let i = 0; i < numNumbers - 1; i++) {
-    const randomNumber = Math.floor(Math.random() * remaining);
-    remaining -= randomNumber;
-    randomNumbers.push(randomNumber);
-  }
-  randomNumbers.push(remaining);
-  return _.shuffle(randomNumbers);
-}
-
-// series and data will always be the same length
-function addToSeriesData(series: ChartSeries[], data: number[]) {
-  for (let i = 0; i < series.length; i++) {
-    series[i].data.push(data[i]);
-  }
-}
-
-function generateSeries(): ChartSeries[] {
-  // order of data points is Female, Male, Other, Female 18-44, Female 45-64, Female 65+, Male 18-44, Male
-  const series: ChartSeries[] = [
-    { name: 'Asian', data: [] },
-    { name: 'Black', data: [] },
-    { name: 'White', data: [] },
-    { name: 'Native American', data: [] },
-    { name: 'Pacific Islander', data: [] },
-  ];
-  // for each of the three gender identity totals,
-  const genderTotals = generateRandomNumbersThatAddUpTo(100, 3);
-  for (const genderTotal of genderTotals) {
-    // generate the race breakdown for the gender totals
-    const genderTotalRaceBreakDown = generateRandomNumbersThatAddUpTo(genderTotal, 5);
-    addToSeriesData(series, genderTotalRaceBreakDown);
-    // get the three age group breakdowns
-    const ageGroupGenderTotal = generateRandomNumbersThatAddUpTo(genderTotal, 3);
-    // get race breakdowns for each of the gender age groups
-    for (const genderAgeGroupTotal of ageGroupGenderTotal) {
-      const genderAgeGroupRaceBreakdown = generateRandomNumbersThatAddUpTo(genderAgeGroupTotal, 5);
-      addToSeriesData(series, genderAgeGroupRaceBreakdown);
-    }
-  }
-  return series;
-}
 export const CohortEditor: React.FC<CohortEditorProps> = (props) => {
   const {
     onStateChange,
@@ -666,41 +602,8 @@ export const CohortEditor: React.FC<CohortEditorProps> = (props) => {
   const [cohort, setCohort] = useState<Cohort>(originalCohort);
   const [snapshotRequestParticipantCount, setSnapshotRequestParticipantCount] =
     useLoadedData<SnapshotBuilderCountResponse>();
-  const [cohortDemographics1] = useState<CohortDemographics>({
-    categories: [
-      { short: 'Female' },
-      { short: 'Male' },
-      { short: 'Other', long: 'Nonbinary, 2 Spirit, Genderqueer, etc.' },
-    ],
-    series: [{ name: 'Overall', data: generateRandomNumbers(3, 100) }],
-    title: 'Gender identity',
-    yTitle: 'AVERAGE AGE',
-    height: '250rem',
-    legendEnabled: false,
-    showSeriesName: false,
-  });
-  const [cohortDemographics2] = useState<CohortDemographics>({
-    categories: [
-      { short: 'Female' },
-      { short: 'Female 18-44' },
-      { short: 'Female 45-64' },
-      { short: 'Female 65+' },
-      { short: 'Male' },
-      { short: 'Male 18-44' },
-      { short: 'Male 45-64' },
-      { short: 'Male 65+' },
-      { short: 'Other', long: 'Nonbinary, 2 Spirit, Genderqueer, etc.' },
-      { short: 'Other 18-44', long: 'Nonbinary, 2 Spirit, Genderqueer, etc. 18-44' },
-      { short: 'Other 45-64', long: 'Nonbinary, 2 Spirit, Genderqueer, etc. 45-64' },
-      { short: 'Other 65+', long: 'Nonbinary, 2 Spirit, Genderqueer, etc. 65+' },
-    ],
-    series: generateSeries(),
-    title: 'Gender identity, current age, race',
-    yTitle: 'OVERALL PERCENTAGE',
-    height: '500rem',
-    legendEnabled: true,
-    showSeriesName: true,
-  });
+  const [cohortAges, setCohortAges] = useState<CohortDemographics>(cohortAgeData);
+  const [cohortDemographics, setCohortDemographics] = useState<CohortDemographics>(cohortDemographicData);
 
   function chartOptions(cohortDemographics: CohortDemographics) {
     return {
@@ -767,6 +670,20 @@ export const CohortEditor: React.FC<CohortEditorProps> = (props) => {
     );
   }, [snapshotId, setSnapshotRequestParticipantCount, cohort]);
 
+  useEffect(() => {
+    setTimeout(() => {
+      cohortAgeData.series = generateAgeSeries(3, 90);
+      setCohortAges(cohortAgeData);
+    }, 2000);
+  }, [snapshotRequestParticipantCount]);
+
+  useEffect(() => {
+    setTimeout(() => {
+      cohortDemographicData.series = generateDemographicSeries();
+      setCohortDemographics(cohortDemographicData);
+    }, 2000);
+  }, [snapshotRequestParticipantCount]);
+
   return div({ style: { display: 'flex' } }, [
     div([
       h(CohortEditorContents, {
@@ -816,8 +733,8 @@ export const CohortEditor: React.FC<CohortEditorProps> = (props) => {
           ? formatCount(snapshotRequestParticipantCount.state.result.total)
           : h(Spinner),
       ]),
-      h(Chart, { options: chartOptions(cohortDemographics1) }),
-      h(Chart, { options: chartOptions(cohortDemographics2) }),
+      h(Chart, { options: chartOptions(cohortAges) }),
+      h(Chart, { options: chartOptions(cohortDemographics) }),
     ]),
   ]);
 };
